refactor: simplify レビュー対応 (共通化 / Fenwick GrowTo O(N) 化 等)
- ComputeVisibleNodeRange を layout_cache.h に共通化、未使用 nodes/theme 引数を削除
- dirty 選定述語を ViewportClip::ShouldMeasure に集約 (RunSerial / RunParallel)
- FloatFenwick::GrowTo を O(N log N) → O(N) (Build の逆操作 + 再 Build)
- parser CurrentTextMatchesRawSlice の probe 短絡 (worst-case で len 全比較) を 1 wmemcmp に簡素化
- small_vector::constexpr_memcpy のメンバテンプレ T シャドウを U に rename
- LowercaseTable::col_count を NodeTableData と同じ uint16_t に統一
- FinalizeOwnedTextLineCount で view_base_ を nullptr リセット (stale ポインタ防止)

+++ src/ui/search_state.h
@@ -134,7 +134,7 @@ private:
         std::pmr::wstring buffer;
         // NodeTableData::cell_text_starts のコピー (size = row_count * col_count + 1)。
         std::pmr::vector<uint32_t> offsets;
-        size_t col_count = 0;
+        uint16_t col_count = 0;
     };
     struct LowercaseCache {
         std::pmr::wstring buffer;                            // 全ノードの lower text を連結
